% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfpca_sim.R
\name{MFPCA_cov}
\alias{MFPCA_cov}
\title{Function to calculate the multivariate FPCA for a given covariance matrix and
univariate basis functions}
\usage{
MFPCA_cov(cov, basis_funs, scores = NULL, weights = NULL)
}
\arguments{
\item{cov}{Covariance matrix of the basis functions coefficients.}

\item{basis_funs}{List with basis functions on each dimension. The basis
functions are funData objects}

\item{scores}{Matrix (n rows, B columns) containing the basis functions
coefficients. Defaults to NULL which does not calculate the multivariate
scores.}

\item{weights}{Vector of weights, defaults to 1 for each element}
}
\description{
Function to calculate the multivariate FPCA for a given covariance matrix and
univariate basis functions
}
