% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjm_bamlss.R
\name{mjm_bamlss}
\alias{mjm_bamlss}
\title{Family for Flexible Multivariate Joint Model}
\usage{
mjm_bamlss(...)
}
\arguments{
\item{...}{All arguments are actually hard coded as needed by the
implementation.}
}
\description{
This function specifies the different predictors and link functions as well
as the corresponding transform/updating/sampling functions as well as the
predict function.
}
