% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helperfun.R
\name{sim_jmb_predict}
\alias{sim_jmb_predict}
\title{Simulation Helper Function - Predict the Results for JMbayes-Models}
\usage{
sim_jmb_predict(m, wd, model_wd, data_wd, rds = TRUE, gamma_timeconst = TRUE)
}
\arguments{
\item{m}{Vector containing all the file names of the models.}

\item{wd}{Path to simulations folder.}

\item{model_wd}{Simulation setting folder where the models are saved.}

\item{data_wd}{Simulation data folder.}

\item{rds}{Objects are saved as .rds files (for backwards compatibility when
.Rdata files were used). Defaults to TRUE.}

\item{gamma_timeconst}{Only implemented for timeconstant gamma predictors. If
FALSE a warning message is returned.}
}
\description{
This function takes all the models listed in a folder and predicts the fit.
}
