% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjm_bamlss.R
\docType{package}
\name{varbinq}
\alias{varbinq}
\title{Flexible Joint Models for Multivariate Longitudinal and Time-to-Event Data}
\description{
This package contains all functions and implementations of the corresponding
paper by Volkmann, Umlauf, Greven: "Flexible joint models for multivariate
longitudinal and time-to-event data using multivariate functional principal
components". Code to reproduce the simulation and analysis as well as
additional information on the model fitting process are contained in
the "inst" folder.
}
