% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcre_smooth.R
\name{Predict.matrix.unc_pcre.random.effect}
\alias{Predict.matrix.unc_pcre.random.effect}
\title{mgcv-style constructor for prediction of PC-basis functional random effects}
\usage{
\method{Predict.matrix}{unc_pcre.random.effect}(object, data)
}
\arguments{
\item{object}{a smooth specification object, see
\code{\link[mgcv]{smooth.construct}}}

\item{data}{see \code{\link[mgcv]{smooth.construct}}}
}
\value{
design matrix for PC-based functional random effects
}
\description{
This predictor function uses time-information saved in the object. This is
handled within the bamlss-transform function, so this function is not
exported.
}
\author{
Alexander Volkmann, adapted from 'Predict.matrix.pcre.random.effect
 by F. Scheipl (adapted from 'Predict.matrix.random.effect' by S.N. Wood).
}
