% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcre_smooth.R
\name{smooth.construct.unc_pcre.smooth.spec}
\alias{smooth.construct.unc_pcre.smooth.spec}
\title{mgcv-style constructor for PC-basis functional random effects (no constraint)}
\usage{
\method{smooth.construct}{unc_pcre.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth specification object, see
\code{\link[mgcv]{smooth.construct}}}

\item{data}{see \code{\link[mgcv]{smooth.construct}}}

\item{knots}{see \code{\link[mgcv]{smooth.construct}}}
}
\value{
An object of class \code{"random.effect"}. See
\code{\link[mgcv]{smooth.construct}}
 for the elements that this object will contain.
}
\description{
Sets up design matrix for functional random effects based on the PC scores
of the covariance operator of the random effect process. Note that there is
no constraint on the smoother.
See \code{\link[mgcv]{smooth.construct.re.smooth.spec}} for more details on
\code{mgcv}-style smoother specification
and \code{\link[refund]{smooth.construct.pcre.smooth.spec}} for the
corresponding \code{refund} implementation.
}
\author{
Alexander Volkmann; adapted from 'pcre' constructor by F. Scheipl
 (adapted from 're' constructor by S.N. Wood).
}
